mental-health-project/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
youtube-transcript-api==0.6.1
pydantic==2.5.0
flask
flask-cors
